+++ dart/lib/src/noop_hub.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
-
 import 'hub.dart';
 import 'protocol.dart';
 import 'sentry_client.dart';
+import 'sentry_user_feedback.dart';
 
 class NoOpHub implements Hub {
   NoOpHub._();
@@ -79,5 +79,10 @@ class NoOpHub implements Hub {
   @override
   Map<String, String> traceHeaders() {
     return {};
+  }
+
+  @override
+  Future<SentryId> captureUserFeedback(SentryUserFeedback userFeedback) async {
+    return SentryId.empty();
   }
 }