+++ flutter/lib/src/file_system_transport.dart
@@ -10,9 +10,9 @@ class FileSystemTransport implements Transport {
   final SentryOptions _options;
 
   @override
-  Future<SentryId> send(SentryEnvelope envelope) async {
+  Future<SentryId?> send(SentryEnvelope envelope) async {
     final envelopeData = <int>[];
-    await envelope.envelopeStream().forEach(envelopeData.addAll);
+    await envelope.envelopeStream(_options).forEach(envelopeData.addAll);
     // https://flutter.dev/docs/development/platform-integration/platform-channels#codec
     final args = [Uint8List.fromList(envelopeData)];
     try {
@@ -27,6 +27,6 @@ class FileSystemTransport implements Transport {
       return SentryId.empty();
     }
 
-    return envelope.header.eventId ?? SentryId.empty();
+    return envelope.header.eventId;
   }
 }